Using Root Hook Plugin
This, according to the documentation, allows testing in parallel
mode.
I've used the array mode here, which allows multiple hooks of a
given type.  It's really not necessary for this project, but I
was playing with the various options and this appears the most
flexable.
I also used arrow funtions instead of regular functions and they
appear to work equally well for the purposes of setting a handfull
of global variables.